removed unused logics from gson builder and renamed builder (53)
