+++ reports/src/main/java/io/jmix/reports/entity/ReportTemplate.java
@@ -24,7 +24,7 @@ import io.jmix.core.entity.annotation.SystemLevel;
 import io.jmix.core.metamodel.annotation.DependsOnProperties;
 import io.jmix.core.metamodel.annotation.InstanceName;
 import io.jmix.core.metamodel.annotation.JmixEntity;
-import io.jmix.reports.app.StrategyGsonBuilder;
+import io.jmix.reports.app.MetadataFieldsIgnoringGsonBuilder;
 import io.jmix.reports.entity.charts.AbstractChartDescription;
 import io.jmix.reports.entity.pivottable.PivotTableDescription;
 import io.jmix.reports.entity.table.TemplateTableDescription;
@@ -358,15 +358,15 @@ public class ReportTemplate implements com.haulmont.yarg.structure.ReportTemplat
         if (getContent() == null)
             return null;
 
-        return new StrategyGsonBuilder()
-                .build(StrategyGsonBuilder.SerializationPolicy.ignoreMetadata)
+        return new MetadataFieldsIgnoringGsonBuilder()
+                .build()
                 .fromJson(new String(getContent(), StandardCharsets.UTF_8), TemplateTableDescription.class);
     }
 
     public void setTemplateTableDescription(TemplateTableDescription description) {
         if (description != null && getReportOutputType() == ReportOutputType.TABLE) {
-            byte[] template = new StrategyGsonBuilder()
-                    .build(StrategyGsonBuilder.SerializationPolicy.ignoreMetadata)
+            byte[] template = new MetadataFieldsIgnoringGsonBuilder()
+                    .build()
                     .toJson(description)
                     .getBytes(StandardCharsets.UTF_8);
